pyproject.toml
[tool.poetry]
name = "activity-monitor"
version = "0.1.0"
description = ""
authors = ["elpachongco <earlsiachongco@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.8"
Flask = "^2.3.2"
setuptools = "^68.0.0"
flasgger = "0.9.7b2"
python-dotenv = "^1.0.0"

[tool.poetry.dev-dependencies]
black = "^23.7.0"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
